Changement double en long dans Circle de CSchema

+++ CSchema.cpp
@@ -274,10 +274,10 @@ bool CSchema::Circle(vector<string> aInst)
     int y = VectStringToInt(aInst [2]);
     int radius = VectStringToInt(aInst [3]);
 
-	double xMax = (double)x+(double)radius;
-	double xMin = (double)x-(double)radius;
-	double yMax = (double)y+(double)radius;
-	double yMin = (double)y-(double)radius;
+	long xMax = (long)x+(long)radius;
+	long xMin = (long)x-(long)radius;
+	long yMax = (long)y+(long)radius;
+	long yMin = (long)y-(long)radius;
 
 	// Gestion du fait que le rayon est en dehors du cadre de travail
 	if (xMax>maxInt || xMin<minInt || yMax>maxInt || yMin<minInt)
